+++ NEWS.md
@@ -1,6 +1,6 @@
 # NEWS
 
-## Polars R Package (development version)
+## Polars R Package 0.16.4
 
 ### New features